+++ src/widget/widgets/button.rs
@@ -3,28 +3,28 @@ use crate::{PeroxideEvent, Widget, widget::Interactable};
 pub struct Button {
     pub parent: Option<Box<dyn Widget>>,
     pub child: Option<Box<dyn Widget>>,
-    pub x: i32,
-    pub y: i32,
-    pub width: i32,
-    pub height: i32,
+    pub x: f64,
+    pub y: f64,
+    pub width: f64,
+    pub height: f64,
     pub color: wgpu::Color,
     pub id: i32,
     pub on_click: Option<Box<dyn Fn(PeroxideEvent)>>,
-    pub decoration: Option<Box<dyn Widget>>,    
+    pub decoration: Option<Box<dyn Widget>>,
 }
 
 impl Widget for Button {
-    fn width(&self) -> i32 {
+    fn width(&self) -> f64 {
         self.width
     }
-    fn x(&self) -> i32 {
+    fn x(&self) -> f64 {
         self.x
     }
-    fn y(&self) -> i32 {
+    fn y(&self) -> f64 {
         self.y
     }
 
-    fn height(&self) -> i32 {
+    fn height(&self) -> f64 {
         self.height
     }
 }
@@ -38,5 +38,3 @@ impl Interactable for Button {
         }
     }
 }
-
-
