+++ src/widget/widgets/str.rs
@@ -1,23 +1,21 @@
 use wgpu::Color;
 
-
-use crate::{window::State, Widget};
+use crate::{Widget, window::State};
 
 impl Widget for &str {
-    fn width(&self) -> i32 {
-        100 
+    fn width(&self) -> f64 {
+        100.0
     }
-    fn height(&self) -> i32 {
-        100 
+    fn height(&self) -> f64 {
+        100.0
     }
     fn color(&self) -> Color {
         Color::WHITE
     }
-    fn x(&self) -> i32 {
-        0 
+    fn x(&self) -> f64 {
+        0.0
     }
-    fn y(&self) -> i32 {
-        0 
+    fn y(&self) -> f64 {
+        0.0
     }
-   
 }
